# src/letter-signature/dto/update-letter-signature.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateLetterSignatureDto } from './create-letter-signature.dto';

export class UpdateLetterSignatureDto extends PartialType(CreateLetterSignatureDto) {}