# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/letter/{id}": {
      "get": {
        "operationId": "LetterController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Letter"
        ]
      },
      "patch": {
        "operationId": "LetterController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLetterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Letter"
        ]
      },
      "delete": {
        "operationId": "LetterController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Letter"
        ]
      }
    },
...
    "/api/letter-signature": {
      "post": {
        "operationId": "LetterSignatureController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLetterSignatureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "LetterSignature"
        ]
      },
      "get": {
        "operationId": "LetterSignatureController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Nomor halaman yang ingin ditampilkan",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Jumlah data per halaman",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Kata kunci untuk pencarian berdasarkan nama atau deskripsi",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterSignature"
        ]
      }
    },
    "/api/letter-signature/{id}": {
      "get": {
        "operationId": "LetterSignatureController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterSignature"
        ]
      },
      "patch": {
        "operationId": "LetterSignatureController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLetterSignatureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterSignature"
        ]
      },
      "delete": {
        "operationId": "LetterSignatureController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "LetterSignature"
        ]
      }
    },
...
    "/api/general-letter-submission": {
      "post": {
        "operationId": "GeneralLetterSubmissionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGeneralLetterSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "GeneralLetterSubmission"
        ]
      },
      "get": {
        "operationId": "GeneralLetterSubmissionController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Nomor halaman yang ingin ditampilkan",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Jumlah data per halaman",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Kata kunci untuk pencarian berdasarkan nama atau deskripsi",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "GeneralLetterSubmission"
        ]
      }
    },
    "/api/general-letter-submission/{id}/print": {
      "get": {
        "operationId": "GeneralLetterSubmissionController_print",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "GeneralLetterSubmission"
        ]
      }
    },
    "/api/general-letter-submission/{id}": {
      "get": {
        "operationId": "GeneralLetterSubmissionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "GeneralLetterSubmission"
        ]
      },
      "patch": {
        "operationId": "GeneralLetterSubmissionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGeneralLetterSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "GeneralLetterSubmission"
        ]
      },
      "delete": {
        "operationId": "GeneralLetterSubmissionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "GeneralLetterSubmission"
        ]
      }
    },
    "/api/document-submission": {
      "post": {
        "operationId": "DocumentSubmissionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentLetterSubmissionId": {
                    "type": "number"
                  },
                  "letterDocumentId": {
                    "type": "number"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "DocumentSubmission"
        ]
      },
      "get": {
        "operationId": "DocumentSubmissionController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Nomor halaman yang ingin ditampilkan",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Jumlah data per halaman",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Kata kunci untuk pencarian berdasarkan nama atau deskripsi",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "DocumentSubmission"
        ]
      }
    },
    "/api/document-submission/{id}": {
      "get": {
        "operationId": "DocumentSubmissionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "DocumentSubmission"
        ]
      },
      "patch": {
        "operationId": "DocumentSubmissionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentLetterSubmissionId": {
                    "type": "number"
                  },
                  "letterDocumentId": {
                    "type": "number"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "DocumentSubmission"
        ]
      },
      "delete": {
        "operationId": "DocumentSubmissionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "DocumentSubmission"
        ]
      }
    }
  },
  "info": {
    "title": "DIGILET",
    "description": "Dokumentasi API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateLetterDto": {
        "type": "object",
        "properties": {
          "letterName": {
            "type": "string",
            "example": "Surat Pengantar",
            "description": "nama surat"
          },
          "referenceNumber": {
            "type": "string",
            "example": "123/456/789",
            "description": "nomor surat"
          },
          "expiredDate": {
            "type": "number",
            "example": "2023-09-10",
            "description": "tanggal kedaluwarsa"
          },
          "letterNumberingStart": {
            "type": "number",
            "example": 1,
            "description": "nomor urut surat"
          },
          "category": {
            "type": "string",
            "description": "kategori surat",
            "enum": [
              "faculty",
              "study_program",
              "university",
              "all"
            ]
          },
          "signatureType": {
            "type": "string",
            "description": "jenis tanda tangan",
            "enum": [
              "barcode",
              "digital"
            ]
          }
        }
      },
...
      "CreateLetterSignatureDto": {
        "type": "object",
        "properties": {
          "studentLetterSubmissionId": {
            "type": "number",
            "description": "ID dari student letter submission"
          },
          "generalLetterSubmissionId": {
            "type": "number",
            "description": "ID dari general letter submission"
          },
          "letterSignatureTemplateId": {
            "type": "number",
            "description": "ID dari letter signature template"
          },
          "signature": {
            "type": "string",
            "description": "Signature (string)"
          }
        },
        "required": [
          "letterSignatureTemplateId",
          "signature"
        ]
      },
      "UpdateLetterSignatureDto": {
        "type": "object",
        "properties": {
          "studentLetterSubmissionId": {
            "type": "number",
            "description": "ID dari student letter submission"
          },
          "generalLetterSubmissionId": {
            "type": "number",
            "description": "ID dari general letter submission"
          },
          "letterSignatureTemplateId": {
            "type": "number",
            "description": "ID dari letter signature template"
          },
          "signature": {
            "type": "string",
            "description": "Signature (string)"
          }
        }
      },
...
      "CreateGeneralLetterSubmissionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "ID dari user"
          },
          "letterId": {
            "type": "number",
            "description": "ID dari letter"
          },
          "letterNumber": {
            "type": "string",
            "description": "Nomor surat"
          }
        },
        "required": [
          "userId",
          "letterId"
        ]
      },
      "UpdateGeneralLetterSubmissionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "ID dari user"
          },
          "letterId": {
            "type": "number",
            "description": "ID dari letter"
          },
          "letterNumber": {
            "type": "string",
            "description": "Nomor surat"
          }
        }
      }
    }
  }
}